apps/data-pipeline/python-cron/requirements.txt
dotenv
polars-lts-cpu
requests
pyodbc
typing
pytz